LayoutTests/fast/regions/cssom/offsetLeft-offsetTop-in-region-float-vert-rl-expected.txt

Test offsetLeft and offsetTop in a named flow with floats in vertical-rl mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS divText1.offsetParent is document.body
PASS divText1.offsetLeft is 700
PASS divText1.offsetTop is 200
PASS divText2.offsetParent is document.body
PASS divText2.offsetLeft is 400
PASS divText2.offsetTop is 200
PASS successfullyParsed is true

TEST COMPLETE
